cleanup: Remove not useful global keys
It turns out that these keys are not helpful because they are rarely
in use. They should not be binded globally to avoid consuming short,
concise keys that may be more efficient if allocated for other helpful regular commands.

package-list-packages was needed in the past because everytime I
installed a new package, I would list all of them first. But this
install behavior is recently replaced by regular straight-use-package
interactive command.

use-package-report is only helpful in measuring emacs init load time
which happens once in a few years. I can create a temporary key for it
in a rare testing time but it shoudn't be by default everytime.

org-mode is actually not used much by my day-to-day editing.

look-up-man-page-at-current-word is rarely used in the past 3 years.