cleanup: Remove no longer useful config
Set default-frame to fullscreen is not needed by default. It can be
activated easily by shortcuts from a window manager or emacs. In fact,
most of the time I sit in front of an external monitor and splitting
screens between emacs and web browser is a more common task.

desktop-save-mode is replaced by psession package. Actually, I
have commented this config for too long that it is worth being commited.